Incorporate feedback from code review
In general, this was just implementing @agitter's suggested changes. One of the questions
that was raised was why I unpinned the version of docutils. I had to double check, but it
looks like pinning the version caused a compatibility issue, specifically with the pip-
installed package `sphinx-rtd-theme`. Rather than unpinning Docutils or bumping it down by
a version, I bumped `sphinx-rtd-theme` up two versions (in both `environment.yaml` and
the `pyproject.toml`) to satisfy the dependency conflict. The change appears not to have
caused issues, but I'm also not aware of how the doc-building mechanism works, so I'm
not sure how to give it a more robust test.