Make SPRAS installable
This commit adds the ability to install spras as a python module using pip,
which is something we'll need for running in remote environments such as on
the OSPool. The installation "instructions" used by pip during building are
laid out in `pyproject.toml`. While I chose to use setup-tools as a builder
backend in this first pass, other builder backends with more robust build
features are available and may eventually be switched to (eg flit, poetry).

Several additional files had to be modified to adhere to python packaging
guidelines, the most notable of which was changing `src` to spras, and referencing
the new spras directory instead of src. Several lines of omicsintegrator2.py were
also modified becasue they were dealing with an absolute path, which made the
workflow un-runnable outside of the repo. I verified that all tests and the entirety of
the snakefile still run both in the conda environment from the repo root (which is
how the tool is currently used) and in a separate directory with spras installed where
the input data, snakefile, and config file have been brought along.

For testing, install spras using `pip install .` in the repo root. If the extra
development packages are needed, spras can be installed with the command
`pip install .[dev]`. Spras can also be installed via `pip install -e .` to
make the installation editable. That is, when installed via this command, you
can continue to edit spras in the repo and have those changes reflected anywhere
you run spras on the machine.

Note that several package versions had to be modified (they're noted in
pyproject.toml) to get past a few versioning issues between conda and pypi. To the
extent of my knowledge and testing, these changes didn't affect the way spras runs.
Famous last words?