Move container framework to algorith run call and fix hash lengths
This commit moves the container framework up a level so that each algorithm
is called with the configured framework instead of determining the framework
internally. The logic behind this is that doing so makes testing much easier;
instead of messing with global configs in the tests, just call the PRM's run
function with a different framework specified.

This also moves prepare_volume to `containers.py` to fix the issue with importing
hash lengths in the function. Arguably, I should have moved prepare_volume into
containers.py in the first place.