Add ability to unpack singularity images
When running nested, unprivileged containers singularity containers, there's a bit
of extra setup we need to perform in order to get things working. In particular,
in order to avoid FUSE requirements, we can "unpack" containers into a directory
after we've converted the docker image to a `.sif`.

This PR adds a new configuration option to allow that unpacking, as well as a SPRAS
runtime container that can be used to demonstrate this behavior in HTCondor. Bundled
along with these changes are a variety of package version updates that I needed to
change to get things working.