Use Apptainer with HTCondor by default and point to SPRAS v0.2.0
Two issues solved with this commit -- first, Apptainer seems to be the way to go
when working with HTCondor, both in and out of the OSPool. Instead of having
instructions that say "if you encounter problem XXX, use Apptainer", this just
tells the user to build the Apptainer image in the first place.

Secondly, Neha encountered an issue while testing HTCondor compatibility where
the container's Snakefile version was incompatible with the version of the
Snakefile being transferred from the AP to the EP. This resulted in a confusing
error message and wouldn't be straight forward to recognize for most people, so
I decided I'd urge the users to build their own containers in the first place.
If they do this, then they can be sure there are no compatibility issues.