Add extra troubleshooting for failed jobs and retry failures up to 5 times
To handle potential failures arising from bad luck in a distributed computing environment,
this PR tries to inform users which types of errors might retriable, and attempts to retry
all errors on the user's behalf up to 5 times. The hope is that this will obscure most of
these transient errors from the user's view.