Split container build action into new file, update build+push action
This commit splits the build + push action into its yaml definition and updates
the build+push action we rely on from v1 to v6. We hadn't updated the action
because there was some previous feature we either relied on, or a breaking
change in a newer release we wanted to avoid. In any case, we can't remember what
the feature was, so it's probably time to give the new stuff a go.

I chose to split this into its own workflow definition because some of the next
improvements I'd like to make involve updating the trigger logic. This could be
done through the same `test-spras.yml`, but is cleaner/easier to read and
understand if done as a standalone.