Remove redundant docker pull from container build GHA job
Previously we did an explicit `docker pull` for each of the containers
we wanted to use as a cache base in the build tests. I don't _think_
this is explicilty necessary, as the v6 docker build+push action has
a default cache-from type of "registry":
https://docs.docker.com/reference/cli/docker/buildx/build/#cache-from

In other words, each build step pulls what it needs to without the
explicit management. Barring a need to pre-load containers for some speed
optimization (I couldn't find any), it seems more prudent to let the
action manage things.