Attempt building containers only when those containers have changed
Previously this test built every container every time someone pushed changes
to a PR. While this was okay, it means that we ended up building containers
even when the actual changes to the codebase had nothing to do with those
containers.

This commit attempts to create a separate build+remove templated action that
uses `git diff` against an input path to determine if a container needs to be
built. Workflow in development!