Add context key, provide more code comments, and fix trailing / in paths
I realized the SPRAS container build wouldn't work with the previous action
because it takes a different context not scoped to the container's Dockerfile
directory. This commit (among a few other small cleanups) adds a `context` input
to the new action so that I can pass the `./` value for the SPRAS container build.