Deduplicate build action runs
The previous action trigger of `on: [push, pull_request]` looked like it was
double dipping -- every time you pushed a new set of commits to a pull request,
the workflow triggered twice.

This instead runs the actions on new pull requests, or on synchronize/reopened
events. The synchronize event occurs when new commits are received.

Additionally, this still runs the test whenever someone pushes to master, which
shouldn't be done...