Add minimal `dsub` framework documentation in config.yaml
As Tony pointed out in the PR review, Google Cloud Life Sciences is being
deprecated, so we're not sure how much longer this will be around. Instead
of going through and giving the `dsub` framework verbose documentation, we
explicitly state it's usage is experimental and warn the user that they
should already know what they're doing if they plan to use it.