Add runtime warning if user selects `dsub` framework
This prints a warning for the user if they select the `dsub` framework. Since
we're unsure whether we'll need to support dsub in the future (as our primary
motivation for using it is being deprecated), this warning should prevent
potential users from getting their hopes up that we have long-term plans for it.