Remove old 'doc/' folder
This commit removes the old 'doc/' folder. It moves the reference SPRAS
overview into docs/_static (the canonical place for static images with Sphinx)
and converts the output file formats into an rst, which I did with:
```bash
pandoc output.md --from markdown --to rst -s -o output.rst
```

My notes say the `doc/README.md` is no longer relevant, so I deleted instead
of moving that file.

One bootstrapping issue -- I updated the repo's main README to point to
some documentation about supported PRMs (it previously pointed to the `doc/`
folder), but this requires using the soon-to-be link for the relevant page
in Read The Docs. We haven't deployed these changes yet, but that page should
exist once we do.