feat: omicsintegrator2 offline support
[I mainly used this issue as an excuse to experiment with a git patch workflow over the forking method that seems to be present in SPRAS]

Forces OI2 to work offline by disabling the unused networkx graph as interactive HTML feature (we already have workflows in SPRAS that enable this).

The network error was coming from `output_networkx_graph_as_interactive_html` dynamically fetching jQuery to insert it as an inline script. This could have also been fixed by providing a `local` folder with all of the scripts, but putting the several hundred-kilobyte scripts on this repository did not seem worth it for an unused feature.