Create peer cgroup for PRM containers and aggregate usage stats
This commit adds the needed bits for the main Python process to create
a peer cgroup (linux only) such that when profiling is enabled, the PRM
containers are run under this cgroup with the `memory.peak` and `cpu.stat`
controllers enabled.

Unfortunately we can't just point Python at some PID, because the PRM
containers launch various processes without reporting the PIDs back to
the originating process. This prevents us from regular inline monitoring.