fix(singularity): gc unpacked files
This is a follow-up to https://github.com/Reed-CompBio/spras/pull/145#issuecomment-1970044310, taking the neutral approach where GC is left up to the user, but in such a way where it is made convenient. This addresses the `allpairs` creating git-commitable artifacts on the cwd.

This path could be made a parameter in the encompassing `run_container_singularity` function, though that seems out of scope for SPRAS itself.